fix!(gateway): move emitter into shard processors (#1070)
When calling `Shard::shutdown`, `Shard::shutdown_resumable`,
`Cluster::down`, or `Cluster::down_resumable`, shards would be stopped
but the events stream returned by shards and clusters wouldn't return
`None`. This is due to the events stream containing a receiver of
events, while shard processors contained a sender. However, shards would
keep a copy of the sender, so while the shard processor would be aborted
and its sender dropped the shard's would not be dropped.

To fix this we can move the sender into shard processors. When the
shard processor is dropped so will the only sender. However, individual
shard instances will now only be able to be started the first time;
`Shard::start` can no longer be called multiple times. If a user shuts
down a shard and wants to start it again they will need to create a new
shard instance.

BREAKING CHANGES: `Shard::start` can now only be called once for the
entire duration of each shard instance.