fix!(http/request): raw request builder
`request::RequestBuilder::raw` was accidentally removed for 0.6; this
adds it back in, hiding the existing fields behind methods due to the
type of one of the public fields being inaccurate in the first place.
This is considered a hotfix.

BREAKING CHANGES: `request::Request` fields are now private and behind
getters, `request::RequestBuilder` and `request::Request` initialization
methods now take route references.

Signed-off-by: Zeyla Hellyer <zeyla@hellyer.dev>