feat(http): option to remember invalid token (#1085)
By default we remember when the HTTP client encounters an invalid token
(via the 401 response status). When one is encountered, all future
requests are short-circuited and return a 401.

This option allows a user to disable this functionality and to always
continue with a request, even if past ones have encountered a 401
response. The new option is
`client::ClientBuilder::remember_invalid_token`.