enhancement(gateway): use ratelimiter over throttling (#1061)
Use a leaky bucket ratelimiter to ratelimit requests instead of using a
throttler to throttle outgoing shard commands to 120 commands per 60
seconds.

The throttler would simply throttle commands to one per 520
milliseconds, causing some availability to be lost and removing the
ability to send burst commands such as member chunk requests.

By using a leaky bucket implementation there is zero delay when tickets
in the ratelimiter are still available. To ensure that heartbeats can
still be sent without delay some availability is reserved for the 
heartbeater. This is compensated for by substracting these from the 120
requests Discord allows to be sent every 60 seconds. In uneven
designated heartbeat intervals there may be one too many tickets
reserved for a given minute.