refactor(gateway/ratelimiter): heartbeat allotting
Rework the logic around how heartbeat tokens are allotted in the
ratelimiter. Instead of doing some casting with a Clippy ignore we can
divrem the reset duration and the new heartbeat interval, taking the
quotient and adding one to be cautious in the event there's a remainder.
This removes the need to perform a lossy cast and removes the Clippy
lint.

Tests have been added.

Relates to #1061.

Blocks on #1101.

Signed-off-by: Zeyla Hellyer <zeyla@hellyer.dev>