fix(http/response): correct request errors (#1100)
The `ResponseFuture`'s underlying HTTP client execution future has a
timeout on it. The `match` branch for how to handle the result of the
execution has three cases:

1. The request being successfully executed and getting an HTTP client
response back
2. The request timing out
3. An error while executing the request

In the handling of cases two and three the error would be transformed
into our own error type, but the handling for it is reversed: if a
request timed out we'd say there was an error executing the request
and vice versa.

This fixes that by reversing the error types in the branches.