enhancement(model/ready): remove option from shard
Change the `twilight_model::gateway::payload::incoming::Ready::shard`
field from being an `Option<[u64; 2]>` to being just a `[u64; 2]`; this
field is always present and does not need to be wrapped in an `Option`.

The field has, additionally, been documented.

BREAKING CHANGES: `Ready::shard` is no longer wrapped in an `Option`.

Signed-off-by: Zeyla Hellyer <zeyla@hellyer.dev>