release(model): bump to 0.6.3
Additions

Add `Command::kind`, which is a `CommandType`, in order to support
Message and User application commands ([#1107] - [@7596ff]).

Add `CommandInteractionDataResolved::messages`, which is a list of
messages relevant to the interaction ([#1107] - [@7596ff]).

Add `MessageType::ContextMenuCommand`, which is relevant to Message and
User commands ([#1107] - [@7596ff]).

Add the `{Current, User}::{accent_color, banner}` fields ([#1127] -
[@zeylahellyer]).

Fixes

`Command` deserialization no longer errors if the `options` field was
missing ([#1112] - [@Erk-]).

[#1107]: https://github.com/twilight-rs/twilight/pull/1107

[@7596ff]: https://github.com/7596ff
[@zeylahellyer]: https://github.com/zeylahellyer