feat(cache,gateway,http,model): threads (#1142)
This PR implements threads. It combines the following PRs: #805, #919, #995, #996, #997, #998, #1015, #1019, #1025, #1027, #1074, #1120, and #1163. It is comprised of work by @Erk-, @rxdn, @zeylahellyer, and myself. 

There is still more to be done, but most of it is cosmetic work. It's better to have support for threads and incrementally improve it than to keep it locked in a feature branch.

The only thing that could be considered breaking is the DIscord API version being updated to 9. This version includes support for threads and nothing else that I am aware of, so I don't think it is actually breaking on our end.

Co-authored-by: Ryan <29165304+rxdn@users.noreply.github.com>
Co-authored-by: Zeyla Hellyer <zeyla@hellyer.dev>
Co-authored-by: Erk <Erk-@users.noreply.github.com>