feat!(model): parse iso 8601 datetime strings
Parse and format Discord's ISO 8601 datetime strings using a custom
parser and formatter. By doing this we can store datetimes as u64s in
the form of Unix timestamps, removing a large number of String
allocations. The parser supports only a strict subset of ISO 8601: dates
separated by dashes with a year, followed by long time with a time
offset of `+00:00` and optional microseconds.

The Discord API provides ISO 8601 datetimes in two variants: with
microseconds and without microseconds. Whether microseconds are included
seems to be split down the middle in frequency; most datetime strings
with microseconds only have millisecond precision while only a small
handful of responses have a true millisecond precision. Due to this the
parser supports the full range of microsecond precision. Datetime
between the Discord epoch (2010) and 2038 are supported.

The formatter works in a similar fashion: support for formatting with
microseconds and without microseconds. The default is to format with
microseconds. This functionality is togglable with a
`TimestampIso8601Display::with_microseconds` configuration method on the
display implementation.

The root `Timestamp` type contains only a non-zero 64-bit unsigned
integer. It can be instantiated in a number of ways: `from_secs` which
takes an amount of seconds from the Unix epoch; `from_micros` which
takes an amount of microseconds; `parse` which parses an input string;
`TryFrom<&str>` to do the same; and `FromStr` to do the same. The
methods are all `const` functions.

`Timestamp` contains two methods for retrieving the inner timestamp:
`as_secs` to retrieve the number of seconds since the Unix epoch and
`as_micros` to retrieve the number of microseconds since the Unix epoch.

`Timestamp` itself does not implement `Display`; it would be ambiguous
whether to display the timestamp in its integer representation or a
formatted ISO 8601 datetime. Because of this, users must go through one
of two methods: the aforementioned `Timestamp::as_secs` and
`Timestamp::as_micros` methods or `Timestamp::iso_8601` to retrieve a
Display formatter to format into an ISO 8601 datetime.

`Timestamp` is used in place of `String`s on fields such as `joined_at`,
`premium_since`, `timestamp`, and so on. Included are
`serde::Deserialize` and `serde::Serialize` implementations on
`Timestamp` itself and a `serde::Serialize` implementation on the
Display formatter. `Timestamp`s contain a `NonZeroU64` for greater
efficiency when wrapped in an `Option`.

The basic API looks like this:

```rust
/// Representation of a Unix timestamp with milliseconds.
pub struct Timestamp(_);

impl Timestamp {
    /// Create a non-zero timestamp from a Unix timestamp with microseconds
    /// precision if the value is not zero.
    pub const fn from_micros(unix_microseconds: u64) -> Option<Self>;

    /// Create a non-zero timestamp from a Unix timestamp with seconds precision
    /// if the value is not zero.
    pub const fn from_secs(unix_seconds: u64) -> Option<Self>;

    /// Parse a timestamp from an ISO 8601 datetime string emitted by Discord.
    pub const fn parse(datetime: &str) -> Result<Self, TimestampParseError>;

    /// Total number of seconds within the timestamp.
    pub const fn as_secs(self) -> u64;

    /// Total number of microseconds within the timestamp.
    pub const fn as_micros(self) -> u64;

    /// Create a Display implementation to format the timestamp as an ISO 8601
    /// datetime.
    pub const fn iso_8601(self) -> TimestampIso8601Display;
}

impl FromStr for Timestamp;

impl<'de> Deserialize<'de> for Timestamp;

impl Serialize for Timestamp;

impl TryFrom<&'_ str> for Timestamp;

/// Display implementation to format a [`Timestamp`] in an ISO 8601 format.
pub struct TimestampIso8601Display { .. }

impl TimestampIso8601Display {
    /// Create a new ISO 8601 display formatter for a timestamp.
    pub(super) const fn new(timestamp: Timestamp) -> Self;

    /// Get the inner timestamp.
    pub const fn get(self) -> Timestamp;

    /// Whether to format the timestamp with microseconds.
    pub const fn with_microseconds(mut self, with_microseconds: bool) -> Self;
}

impl Display for TimestampIso8601Display;

impl Serialize for TimestampIso8601Display;
```

Example of parsing a datetime string and formatting it back into the
input:

```rust
let input = "2021-03-16T14:29:19.046000+00:00";

assert_eq!(input, Timestamp::from_str(input)?.iso_8601().to_string());
```

Example of creating a timestamp from its number of microseconds and
formatting it into an ISO 8601 datetime string without microseconds:

```rust
let timestamp = Timestamp::from_micros(1_580_608_922_020_000)?;

assert_eq!(
    "2020-02-02T02:02:02+00:00",
    timestamp.iso_8601().with_microseconds(false).to_string(),
);
```

BREAKING CHANGES: datetime fields stored as `String`s are now stored as
`twilight_model::datetime::Timestamp`s.

Signed-off-by: Zeyla Hellyer <zeyla@hellyer.dev>