release(standby): bump to 0.7.0
Changes

`Standby` no longer implements `Clone`, because it is no longer
internally wrapped in an `Arc` ([#1067] - [@zeylahellyer]). To retain
this functionality, you can wrap them it an `Arc` or a `Rc` manually.

Refactor and fully document internals ([#1159] - [@zeylahellyer]).
Internal logic and functionality is largely untouched.

The MSRV has been updated to 1.53 ([#1161] - [@7596ff]).

[@7596ff]: https://github.com/7596ff
[@zeylahellyer]: https://github.com/zeylahellyer

[#1067]: https://github.com/twilight-rs/twilight/pull/1067
[#1159]: https://github.com/twilight-rs/twilight/pull/1159
[#1161]: https://github.com/twilight-rs/twilight/pull/1161