chore(examples): fold in webhook-slash
Adds the contents of `twilight-rs/examples` to the new examples
directory.

If you are having rust-analyzer false positives, it is because
ed25519-dalek sets a configuration option that hides its entire contents
from its public API while in test mode. Since rust-analyzer compiles in
test mode, this causes the false positive. This will be fixed in the 1.1
release of that crate.

Add this line to VS Code settings, or equivalently configure your
rust-analyzer settings:

```
    "rust-analyzer.cargo.unsetTest": [
        "ed25519-dalek"
    ],
```