fix(model): account for the `type` of `CommandDataOption` in deserialization (#1240)
Because the JSON doubles sent by discord sometimes seems to be without
trailing decimals (eg `5.0` was sent as `5`) it would sometimes cause
the deserialization to fail. This PR rewrites the deserialization of
`CommandDataOption` to look at the type of the value when
deserialization happens so it will convert `5` to a double if the type
is a `Number`.