fix(model/datetime): double leap year handling
Our ISO8601 parser erroneously had a check for whether the year was a
leap year and if the month was after February, increasing the day by 1
if so to account for February 29th. However, in another part of the
parser, this is already handled by a call to `Month::running_total`,
which takes the leap year into account. This caused dates like
"2016-03-01T07:33:19.000000+00:00" to be a day later than they should
have been.

Signed-off-by: Zeyla Hellyer <zeyla@hellyer.dev>