feat(standby)!: expose results of event processing (#1160)
Expose the results of processing an event by keeping a counter of what
results we get from bystander checks. These statistics are:

- **dropped**: the channel was closed due to the receiver end dropping
  their half
- **fulfilled**: a future bystander was matched and an event sent to it
- **sent**: a stream bystander was matched and an event sent to it

Statistics about bystanders being skipped due to not matching are not
counted because not all bystanders are iterated over.

A caveat with the "fulfilled" and "sent" counters is that it can't be
known whether the receiver half received the event, only whether the
event was sent over the channel. This is thoroughly documented.

Co-authored-by: Cassandra McCarthy <cassie@7596ff.com>