feat(http): support trust-dns (#1310)
This adds back [trust-dns] support to twilight-http. It used to be
possible to use this via reqwest's trust-dns feature flag and was no
longer possible after the switch to hyper.

trust-dns is desirable because it replaces the default [GaiResolver] in
hyper, which uses a blocking getaddrinfo call and runs that in a tokio
threadpool. trust-dns instead provides a fully async DNS resolver on
application level.

trust-dns also provides advanced features like DNS-over-HTTPS,
DNS-over-TLS and DNSSEC, which can be enabled by adding e.g. this to
Cargo.toml: `hyper-trust-dns = { version = "0.3", default-features =
false, features = ["dns-over-rustls", "dnssec-ring"] }`

The crate that is used to do this is a very thin wrapper around
trust-dns, it is a [100 LOC] wrapper to implement `Service<Name>` for
trust-dns' TokioAsyncResolver.  The maintenance effort is therefore
pretty low and porting to newer versions of the dependencies is very
straightforward.

[100 LOC]: https://github.com/Gelbpunkt/hyper-trust-dns/blob/main/src/lib.rs
[GaiResolver]: https://docs.rs/hyper/latest/hyper/client/connect/dns/struct.GaiResolver.html
[trust-dns]: https://docs.rs/trust-dns-resolver/latest/trust_dns_resolver/