refactor(http): add FormBuilder and clean up attachments
This is a proof-of-concept for a new way to manage attachments. Only one
request is completed due to time constraints. It is a first step in
supporting attached files in interaction responses that use
`CallbackData`, and it cleans up and simplifies the `Form` struct as it
was quite unmaintainable and hard to read.