refactor(http,validate)!: move validation to crate (#1331)
This PR refactors much of `http`'s validation to its own crate. This
crate exposes functions and error types that ensure requests are
validated. `http` depends on `validate` and returns `validate`'s error
types from `http`'s own functions.

Besides the names and return types of functions, there are not many
breaking changes. But, this PR is the first step that enables other
changes, such as #1165 and #1298.