Use macro for `ErrorCode` variants
Adding a new `ErrorCode` variant was previously a tedious process due to
having to duplicate the display implementation as a rustdoc comment and
matching `ErrorCode::num()`'s output to `From<u64>`'s input.

To assure this a tests were added (`assert_error_code()`), this is no
longer necessary as the macro correctly handles everything.