ci: automatically label pull requests (#1393)
Add a new GitHub Actions workflow that uses `fuxingloh/multi-labeler` to
automatically label pull requests depending on the file changes and the
PR title.

`c-*` and `t-github-actions` are added depending on the files changed by
the PR. These labels are synced (except `c-all` and `t-github-actions`)
so they will be automatically removed if the PR does no longer change
these files.

`t-*` and `m-breaking-change` are added depending on the PR title. This
will only work if the title respects the Conventional Commits
specification.