feat(gateway): make parse errors much more visible
Bumping it up to error for visibility on both the general error and the payload. This allows easier tracking down of broken payloads even in production environments where the debug level is way too verbose when you are looking for a single broken payload type without knowing what event type is failing.