deps(cache,lavalink,standby): downgrade to dashmap 4 (#1434)
Using `Ref::value` with `dashmap` 5 may produce a reference that outlive
the Ref and thus cause undefined behavior. Issues with the cache have
been reported [on Discord] and *may* be related to that. See
https://github.com/xacrimon/dashmap/issues/167 for more information.

I think we should downgrade dashmap until a patch is released, as
`Ref::value` is used in some places with the cache. The issue is not
affecting dashmap 4.

This reverts https://github.com/twilight-rs/twilight/pull/1336.

[on Discord]: https://discord.com/channels/745809834183753828/745811216579952680/929856218120462396