refactor(cache, model)!: make channel a struct
Convert the `twilight_model::channel::Channel` type from an enum with
many sub-types to one single concise struct.

As it currently stands, our `Channel` type is an enum with three
variants: `Group`, `PrivateChannel`, and `GuildChannel`. The former two
are self-descriptive and structs of only data that should be available
in those cases. `GuildChannel` is another enum with *seven* variants,
three of which are thread types.

One issue with doing this is that the Discord API does not make any
guarantees about what fields will be available in what situations. Put
in another way, a channel is a single resource with no modeled
constraints based on its channel type. This ends up with a lot of
duplication between channels for both us and the end user, as well as
some difficulties with model channel types when new fields are
introduced that don't necessarily make sense in the context of the
individual channel mapping.

We took a few approaches towards this API. Other options we considered
were: merging *some* types while maintaining the enum pattern; grouping
together related guild channel fields into optional groups based on
whether they were available; and flattening all channel domains (groups,
private channels, and guild channels) into one struct with grouped
optional fields. We chose to simply flatten it down all the way, which
brings us to an equivalent implementation as that of Discord's model and
will prevent breaking changes and API oddities in the future. For more
information on these approaches, see issue #1367.

Closes #1367.

Signed-off-by: Zeyla Hellyer <zeyla@hellyer.dev>