feat(http,model)!: support attachments in create_response
This PR adds attachment support to the initial response to an
interaction. Previously, users needed to defer an update and edit in
attachments with another request. However, that might still be
preferable due to the three second response requirement.

It also refactors the `InteractionResponse` struct into something that
more closely matches what is sent to Discord. This struct has been moved
to a new `model` module, `http`. This will eventually contain versions
of models that are only ever sent to Discord.