refactor(model)!: revert moving snowflake to model
Revert "refactor: add snowflake getters to `Id<T: Snowflake>` (#1332)",
effectively moving the `Snowflake` trait back to `util`. The Snowflake
trait isn't a model, it's a helper for some computations; the model
crate otherwise has no other traits other than this one, and so it's
unprecedented and its addition was a mistake.

This reverts commit 540ee9a8e9e4127c1744692a49b50dbac99c2d7d.