refactor(cache/in-memory)!: cached voice state (#1491)
We have a type, `model::CachedVoiceState`, that is a slimmed-down
version of a voice state. However, we don't actually use this type
anywhere -- instead, we cache `twilight_model` voice states. This is an
error from back when the cache was made. This makes the cache use the
cached voice state type.