fix(cache-inmemory): proper handling and cleanup on `UnavailableGuild` (#1506)
Added a test to showcase what the behavior looks like now: Guild gets
marked as unavailable, guild entities are removed from the cache.

Previously, the guild was removed from the cache, but entities remained
(especially in `guild_members` and `members`). When the guild went
available again, it was added to the cache, but `guild_members` was
cleared and all member inserts were skipped due to the members already
being in `members` cache.