perf(cache): remove iter when caching stickers
`InMemoryCache::cache_stickers` operates by first retrieving a mutable
reference to the set of sticker IDs for a guild and then allocating a
Vec of the removed delta of sticker IDs between the cached set and the
new set of IDs. The list of removed sticker IDs is then iterated over to
remove it from the guild's set of IDs, and then iterated over again to
remove the stickers from the cache.

We can instead call `HashSet::retain` to mutate over the set of guild
stickers and remove from the set as necessary, removing from the cached
stickers map at the same time if not retaining.

This removes two iterations (filtering removed IDs and one of the two
actual removal iterations) and an allocation.

Signed-off-by: Zeyla Hellyer <zeyla@hellyer.dev>