feat(cache): add missing cached model fields
Update the in-memory cache models with new fields from their
`twilight_model` equivalents. To ensure that these models stay up to
date in the future, initializers for most models have been extracted
from being in the business logic of the library to being initializer
functions on the cached model types themselves. To go along with this,
initializers now exhaustively destructure models' fields to ensure that
when a new field is added to the `twilight_model` variant of a cached
model, it must be handled in the cached model's initializer as well.

The following fields have been added:

CachedGuild:

- `CachedGuild::max_video_channel_users`
- `CachedMessage::application_id`
- `CachedMessage::components`
- `CachedMessage::interaction`
- `CachedMessage::thread_id`

Field and implementation tests for all cached models have also been
added.

Closes #1555.

Signed-off-by: Zeyla Hellyer <zeyla@hellyer.dev>