feat(validate): button styles and required fields
Validate the required fields on buttons. In particular, this adds two
new error variants: `ButtonConflict`, which occurs when a button has
both a custom ID and URL; and `ButtonStyle`, which occurs when a button
that has a style of `Link` does not have a URL or when a button that
does not have a style of `Link` does not have a custom ID.

New validation in `component_inner` has been added, along with the two
new error variants and tests. This also fixes two invalid portions of
existing tests.

Signed-off-by: Zeyla Hellyer <zeyla@hellyer.dev>