refactor(gateway)!: remove option around gateway_url
By fetching the gateway_url in `ShardBuilder::build` instead of in
`Shard::start`.

This also reduces the number of HTTP requests on Clusters from N shards
to 1 (as it would fetch the gateway_url for each shard on startup).
Accepting and storing a `Client` is therefore needless(???), but it's
kept for now