fix(model)!: multiple autocomplete enabled options - removed `ApplicationCommandAutocomplete` - added `CommandOptionValue::Focused(String, CommandOptionType)`
`ApplicationCommandAutocomplete` has been added in order to fix the deserialization of autocomplete interactions. The issue with this approach is that when a command has more (non String) options the deserialization fails.
This is another way of solving this issue. Instead of having a separate struct the `CommandOptionValue::Focused(String, CommandOptionType)` variant has been added. Which gets used when an autocomplete enabled option is focused.
The reason of removing `ApplicationCommandAutocomplete` is to prevent repeating code.