fix(http,model)!: properly serialize attachments in `CreateResponse`
Previously, `Attachment` structs sent inside `InteractionResponseData`
would both serialize the entire file in the JSON body and fail to
include its ID, which made it inefficient and impossible for Discord to
associate attachment metadata with attachment data.

As a result, users are now required to specify their own ID when adding
attachments to all create-message-like request builders. While breaking
and annoying, this is a more robust solution than others, which would
require a lot more code to temporarily fix a problem.