fix(cache): rework `parent_overwrites`
In the permission calculator extension, the `parent_overwrites` function
was meant to combine a thread's permission overwrites with its parent
channel's overwrites. However, this function's previous behavior was to
retrieve itself from the cache again, and combine its overwrites with
itself. This PR fixes this by actually retrieving the channel from the
`parent_id` field, if it exists, and combining the overwrites in a
clearer way.