refactor(model): make application info generalized
`twilight_model::oauth::CurrentApplicationInfo` isn't only a return type
for `twilight_http`'s `GetUserApplication` request, and is actually used
in more places throughout the API that we do not yet support. The
returned application is just an application - nothing unique about it to
the `GetUserApplication` - so we can rename it to just `Application`.
Previous exports have been deprecated.

Signed-off-by: Zeyla Hellyer <zeyla@hellyer.dev>