refactor!: make tracing required (#1684)
Support for the `tracing` crate as a method for logging is currently
opt-in via respective `tracing` features on all crates that support it.
However, `tracing` is used very widely through the ecosystem, and as a
result we already have a hard downstream dependency on it. We can clean
up our codebase, our feature lists, and documentation by removing the
feature and making it a hard dependency.

Users can still opt in to or out of logging through `tracing`'s
configuration.