refactor(gateway): allow ptr_arg on conditional
Allow the `clippy::ptr_arg` lint on `compression::add_url_feature`,
which conditionally uses the provided `&mut String`. Clippy recommends
making this take `&mut str` where zlib is disabled since it's not used,
but because we'll always be passing an allocated string in anyway this
isn't worth breaking the function up to fix.

Signed-off-by: Zeyla Hellyer <zeyla@hellyer.dev>