feat: inline book into repo (#1798)
Inline the twilight documentation book hosted at our
`twilight-rs.github.io` repository into the main repository `twilight`.
This allows us to keep the book updated much more easily (at the same
time as things are done!), and by being in the main repository it can
have more attention through notifications and being in more of a
spotlight.

Unfortunately, because of how GitHub Pages works we would not have been
able to easily maintain having it located at the root of twilight.rs,
which is a requirement due to it being our main entry point for our
website. Because of this, the next best we can do is switch to
Cloudflare's Pages. We already use for the API documentation and for
DNS, so this is an easy switch over.

A live hosted version of documentation built from these changes can be
found at <https://twilight-book.pages.dev>.