refactor!(gateway): restructure the crate
Refactor and restructure the `twilight-gateway` crate from the ground
up, providing a simpler implementation that is easier to follow the
implementation details of and easier to use.

There are two primary routes this PR accomplishes to achieve these
goals: removing the cluster and removing the shard processor. The
cluster largely duplicated much of the shard and significantly increased
the API surface of the crate. Instead of duplicating the base shard
types we can make them more flexible to allow for mass creation of
shards, which is what the cluster achieved. Meanwhile, the shard
processor was a background task that actually "ran" the logic and state
of the shard and its Websocket connections, so by removing it we can
de-duplicate API between the previous shard and shard processor into
just the shard.

The API is dramatically simplifies with these changes. In summary, this
brings the API down to a shard, its configuration, event type flags,
message ratelimiter, and Websocket messages, streams for initializing
multiple shards, along with some small supporting types. In comparison
to the previous API, the previous API included mostly all this and
nearly double the types to support a cluster half of the API.

A bonus benefit of this is that everything* has been refactored. The
internals of shards are now more understandable, cruft that has
collected over the years has been removed, and the flow of the code is
easier to understand for those attempting to learn it. Internal
documentation, both explicit and self-documented, has been greatly
improved along with this.

This refactor and restructure is based on discussion from Discord and
from discussion #1639.