refactor(model)!: remove `Number`
Deriving `Eq` for `Number` in it's current implementation is just plain
wrong.
As `Eq` is not necessary for these types this is fine. `Eq` is mostly
for `HashMap` keys, which is not a usecase for any of these types.

Why not fix `Number` so that it's `Eq` and `Hash` implementation is
correct? Because floating point numbers are notoriously hard to deal
with.