refactor!(gateway-queue)!: drop TLS features
`LargeBotQueue` requires a `Client` from twilight-http, meaning that
users must also depend on twilight-http to use it.
TLS requirements should therefore only be specified in twilight-http,
with twilight-gateway-queue inheriting said requirements.

Also aligns the feature selection to the README features section.

This PR finally fixes all errors to gateway-queue introduced in #1690
and is 100% the same as #1769