fix(gateway)!: within a stream, reinsert shard on error
Previously, if a shard within a stream encountered any error, not just a
failure, it would never re-insert back into the futures list because it
never executed its drop implementation. This change always returns the
shard when polling `ShardEventStream` or `ShardMessageStream`,
regardless of if an error occurred, which ensures it is re-inserted.

Co-authored-by: Zeyla Hellyer <zeyla@hellyer.dev>