ci: Install stable Rust for all jobs
The GitHub runners come preinstalled with the most recent Rust
version at build time (generated once a week), but this does not
always line up with the latest Rust release. For instance, Rust
1.64 is released 2022-09-22 at 12:00 UTC, but the GitHub runner
was also releaced at that day, but before the Rust release, the
runner will therefore trail behind one version until 2022-09-29
when it's updated. This conflicts with the Twilight release schedule,
(0.14 is scheduled for release 2022-09-29) as CI must pass
(and CI requires 1.64) for PRs to be mereged. So to not delay Twilight's
release, CI's Rust version must be bumped earlier than GitHub does.