fix(gateway): properly handle disconnects (#1932)
This PR moves the session identify queueing from `Shard::with_config()`
to `Shard::identify()`, calls `Shard::identify()` after having received
a `Hello` event when missing a session, and sends a `Resume` event if
having an active session when reconnecting (else it reidentifies).

## Resuming
[Resuming] should be attempted in the following scenarios:
1. Received a reconnect event
   ```rs
       Ok(OpCode::Reconnect) => {
           self.disconnect(Disconnected::Resume);
       }
   ```
2. Disconnected with a close code (close frame contains a close code)
   indicating it can reconnect (`CloseCode::can_reconnect`)
   ```rs 
       Message::Close(maybe_frame) => { 
           self.status = ConnectionStatus::from_close_frame(maybe_frame.as_ref()); 
  
           return Ok(Message::Close(maybe_frame));
   ```
3. Disconnected without receiving a close code (same as above)
4. Received an `InvalidSession` containing `true` (`false` calls
   `Shard::disconnect` with `InvalidateSession` to reidentify, `true`
   with `Resume`)
   ```rs 
       Ok(OpCode::InvalidSession) => { 
           let event = Self::parse_event(buffer)?; 
           self.disconnect(Disconnect::from_resumable(event.data)); 
       }
   ```

`Shard::disconnect()` sets `Shard::status` to `Disconnected` and drops
the `Session` if called with `Disconnect::InvalidateSession`. The next
call to `Shard::next_message` will then call `Shard::reconnect`, which
in most cases will send a `Resume` event as the `Session` will be
present, therefore resuming the session.

The shard will receive a `Hello` event when connecting/reconnecting to
the gateway, it will then identify itself if it does not have an active
`Session` at this point (which it will if it's resuming).

Resuming is *not* the same as reconnecting, resuming will fail if the
shard has missed to many events, in this case it will receive a
`InvalidSession` containing `false` (see resuming scenario 4).

[Resuming]: https://discord.com/developers/docs/topics/gateway#resuming 

## Implementation

A background task is used for awaiting permission from the gateway queue
as the other options for not blocking the shard on long queues are not
ideal:

1. Creating a `Queue::request` future when calling `NextMessageFuture`
   is not possible as the future is not cancel safe
2. Saving the future is irritating as it's not `Debug` nor `Sync`,
   requires an additional field on `Shard`, and requires extending
   `NextMessageFuture`.

Spawning a future is, by contrast, very easy and the logic in
`MessageSender` allows the shard to easily be informed when it's time to
identify.

One could say that identify requests should take priority over other
events, which `MessageSender` doesn't allow, but most events are not
allowed to be sent to the gateway before a session is initialized either
way, so it'd be wrong for user's to queue up events either way before
being identified. The background task will also not cause issues if the
shard got dropped during the queue wait time.